requirements.txt
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
colorlog>=6.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
